requirements.txt
pandas==2.2.2
numpy
matplotlib==3.9.0
scipy==1.13.1
pykrige==1.7.1
rasterio==1.3.10
pyproj==3.6.1
geopandas==0.14.4     # необязательно, но оставим для роста
shapely==2.0.4
pyyaml==6.0.2
